entity  수정

+++ src/main/java/com/example/demo/user/entity/Anime.java
@@ -18,7 +18,7 @@ public class Anime {
     private Long id;
 
     @Column(name="mal_id",unique = true)
-    private Integer malId;
+    private Long malId;
 
     private String title;
